split daos into two, one for logic, one for the ui
